♻️ Move DisplayArea types and utilities for ERDRenderer, including auto layout computation and table node checks
