+++ .changeset/pink-dolls-type.md
@@ -1,0 +1,5 @@
+---
+"@liam-hq/erd-core": patch
+---
+
+♻️ Move DisplayArea types and utilities for ERDRenderer, including auto layout computation and table node checks
