+++ frontend/packages/erd-core/src/components/ERDRenderer/components/ERDContent/index.ts
@@ -1,4 +1,1 @@
 export * from './ERDContent'
-export * from './TableNode'
-export * from './useTableSelection'
-export * from './computeAutoLayout'
