+++ frontend/packages/erd-core/src/components/ERDRenderer/components/LeftPane/LeftPane.tsx
@@ -1,3 +1,4 @@
+import { isTableNode } from '@/features/reactflow/utils'
 import { useVersion } from '@/providers'
 import {
   BookText,
@@ -17,7 +18,6 @@ import {
 } from '@liam-hq/ui'
 import { useNodes } from '@xyflow/react'
 import { useMemo } from 'react'
-import { isTableNode } from '../ERDContent'
 import { CopyLinkButton } from './CopyLinkButton'
 import styles from './LeftPane.module.css'
 import { MenuItemLink, type Props as MenuItemLinkProps } from './MenuItemLink'
