+++ frontend/packages/erd-core/src/components/ERDRenderer/hooks/useTableSelection/index.ts
@@ -1,0 +1,1 @@
+export * from './useTableSelection'
