+++ frontend/packages/erd-core/src/components/ERDRenderer/hooks/useTableSelection/useTableSelection.ts
@@ -2,7 +2,7 @@ import { useCustomReactflow } from '@/features/reactflow/hooks'
 import { highlightNodesAndEdges } from '@/features/reactflow/utils'
 import { updateActiveTableName } from '@/stores'
 import { useCallback } from 'react'
-import type { DisplayArea } from '../types'
+import type { DisplayArea } from '../../types'
 
 type SelectTableParams = {
   tableId: string
