+++ frontend/packages/erd-core/src/components/ERDRenderer/utils/computeAutoLayout/getElkLayout.ts
