+++ frontend/packages/erd-core/src/components/ERDRenderer/utils/computeAutoLayout/index.ts
@@ -1,0 +1,1 @@
+export * from './computeAutoLayout'
