+++ frontend/packages/erd-core/src/features/reactflow/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './columnHandleId'
 export * from './convertDBStructureToNodes'
 export * from './highlightNodesAndEdges'
+export * from './isTableNode'
