+++ frontend/packages/erd-core/src/features/reactflow/utils/isTableNode.ts
@@ -1,0 +1,5 @@
+import type { Node } from '@xyflow/react'
+import type { TableNodeType } from '../types'
+
+export const isTableNode = (node: Node): node is TableNodeType =>
+  node.type === 'table'
